% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adduct.R
\name{adduct}
\alias{adduct}
\title{Common adducts}
\usage{
adduct(F, mode = c("+", "-"))
}
\arguments{
\item{F}{chemical formula, must be written in capital letters}

\item{mode}{ionization mode, either positive '+' or negative '-'}
}
\description{
calculate common adduct ions in pos or neg ion mode
}
\examples{
adduct('C1H4',mode = '-')
adduct('C1H4',mode = '+')
}
