% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Get the compound information}
\usage{
describe(chem, representation = "formula", info = FALSE)
}
\arguments{
\item{chem, }{chemical name of the compound}

\item{representation, }{representation methods, formula is default}

\item{info, }{extra molecular information that users can query}
}
\description{
get compound formula and structure from https://cactus.nci.nih.gov/chemical/structure
}
\examples{
\dontrun{
describe('malic acid', "formula")
describe(c('malic acid', 'citric acid', 'tartaric acid'), "smiles")
}
}
\author{
Yonghui Dong
}
