% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element_Isotope.R
\name{E_iso}
\alias{E_iso}
\title{Element isotopes}
\usage{
E_iso(S)
}
\arguments{
\item{S}{element, can be element symbol (i.e. C) or full name (i.e. Carbon).
Element symbol is case sensitive, which full name is case insensitive.}
}
\description{
Check element isotope information
}
\examples{
 E_iso('C') # element symbol
 E_iso('Carbon') # element full name
 E_iso('carBon') # element full name
}
