% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFC.R
\name{getFC}
\alias{getFC}
\title{calculate fold change}
\usage{
getFC(x, Group = NULL)
}
\arguments{
\item{x}{sample ion intensity matrix, row sample, column feature.}

\item{Group}{sample group information}
}
\value{
a dataframe with mean values and fold changes
}
\description{
calculate fold change among different samples.
}
\examples{
dat <- matrix(runif(2*300), ncol = 2, nrow = 300)
myGroup <- rep_len(LETTERS[1:2], 300)
ret <- getFC(dat, Group = myGroup)
}
