% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMax.R
\name{getMax}
\alias{getMax}
\title{Get the sample name which has the max ion intensity}
\usage{
getMax(x)
}
\arguments{
\item{x}{sample ion intensity matrix, row sample, column feature.}
}
\value{
a data frame
}
\description{
get the sample name which has the max ion intensity
}
\examples{
dat <- cbind.data.frame(mz = c(100, 101, 300), mz2 = c(0, 0 , 1), mz3 = c(1, 9, 1))
rownames(dat) <- c("A", "B", "C")
out <- getMax(dat)
}
