% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPLSDA.R
\name{getOPLSDA}
\alias{getOPLSDA}
\title{calculate VIP values}
\usage{
getOPLSDA(x, Group = NULL)
}
\arguments{
\item{x}{sample ion intensity matrix, row sample, column feature.}

\item{Group}{sample group information}
}
\value{
a dataframe with vip values
}
\description{
calculate VIP values using among different samples groups using OPLSDA.
}
\examples{
dat <- matrix(runif(2*300), ncol = 2, nrow = 300)
myGroup <- rep_len(LETTERS[1:3], 300)
out <- getOPLSDA(dat, Group = myGroup)
}
