% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP.R
\name{getP}
\alias{getP}
\title{get p-values}
\usage{
getP(x, Group = NULL)
}
\arguments{
\item{x}{sample ion intensity matrix, row sample, column feature.}

\item{Group}{sample group information}
}
\value{
a data frame
}
\description{
get p-values from Post Hoc analysis
}
\examples{
dat <- matrix(runif(2*300), ncol = 2, nrow = 300)
myGroup <- rep_len(LETTERS[1:3], 300)
out <- getP(dat, Group = myGroup)
}
