% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iso_mass.R
\name{Iso_mass}
\alias{Iso_mass}
\title{Isotope labelled molecular mass}
\usage{
Iso_mass(F, iso)
}
\arguments{
\item{F, }{chemical formula, case insensitive}

\item{iso, }{labelled elements, case insensitive}
}
\description{
Calculate isotope labelled molecular mass
}
\examples{
Iso_mass(F = 'C7H6O4', iso = '[13]C2[2]H3') # Two 13C and three 2H are labled
}
\author{
Yonghui Dong
}
