% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iso_mz.R
\name{Iso_mz}
\alias{Iso_mz}
\title{Isotope labelled molecular mass}
\usage{
Iso_mz(F, iso, z)
}
\arguments{
\item{F, }{chemical formula, case insensitive}

\item{iso, }{labelled elements, case insensitive}

\item{z}{charge}
}
\description{
Calculate isotope labelled m/z
}
\examples{
Iso_mz(F = 'C7H6O4', iso = '[13]C2[2]H3', z = -1) # Two 13C and three 2H are labled
}
\author{
Yonghui Dong
}
