% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{molecular mass}
\usage{
mass(F)
}
\arguments{
\item{F}{chemical formula, case insensitive}
}
\description{
calculate accurate molecular mass
}
\examples{
 mass('C7H6O4')
 mass('c7H6O4') # case insensitive
 mass(c('K1', 'C5H8', 'nA20')) # vector input
}
\author{
Yonghui Dong
}
