% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_volcano.R
\name{myvolcano}
\alias{myvolcano}
\title{volcano plot}
\usage{
myvolcano(result, compare_group, FC = 2, p = 0.05, show_plot = "iden")
}
\arguments{
\item{result}{result from what2() function}

\item{compare_group}{which groups you want to compare, i.e. c("WT", "Treat1"), only two groups are allowed}

\item{FC}{select fold change values, default = 2}

\item{p}{select p value, default = 0.05}

\item{show_plot}{volcano plot should be plotted according to "iden", identified compounds or "non_iden", non identified compounds. default = "iden"}
}
\description{
volcano plot.
}
\examples{
\dontrun{
myvolcano(result, compare_group = c("WT", "JA"))
}
}
\author{
Yonghui Dong
}
