% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz}
\alias{mz}
\title{Calculate accurate mass to charge ratio}
\usage{
mz(m, z, caseSensitive = FALSE)
}
\arguments{
\item{m}{chemical formula of an ion, case insensitive}

\item{z}{charge}

\item{caseSensitive}{if case sensitive is `False`, the elements are seperated by numbers.
for instance, you can write 'c7h7O1', and}
}
\description{
Calculate accurate mass to charge ratio (m/z)
}
\examples{
 mz('C7H7O1', z = 1)
 mz('C10H6Cl1', z = -1)
 mz('C7h7O4', z = 1) # case insensitive
 mz(c('C7H7O4', 'c1'), z = -1) # vector input
}
\author{
Yonghui Dong
}
