% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{mass accuracy}
\usage{
ppm(m, t, lazy = TRUE)
}
\arguments{
\item{m}{measured m/z}

\item{t}{theoretical m/z}

\item{lazy}{if lazy input is allowed}
}
\description{
calculate the mass accuracy of measured m/z. lazy input allowed
}
\examples{
 ppm(155.03383, 155.03388) # with m/z value
 ppm(155.03383, .03388) # lazy input when the integer parts of m and t are the same
 ppm(155.03384, mz('C7H7O4', z = 1)) # with ion formula
}
\author{
Yonghui Dong
}
