% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tanimoto.R
\name{tanimoto}
\alias{tanimoto}
\title{tanimoto similarity}
\usage{
tanimoto(x)
}
\arguments{
\item{x}{similary}
}
\description{
calculate tanimoto similarities of different compounds
}
\examples{
 x <- data.frame(Samp1=c(0,0,0,1,1,1,0,0,1), Samp2=c(1,1,1,1,1,1,0,0,1))
 tanimoto(x)
}
