% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewTIC.R
\name{viewTIC}
\alias{viewTIC}
\title{view TIC}
\usage{
viewTIC(x, Group = NULL, Batch = NULL, Seq = NULL, Trans = "none")
}
\arguments{
\item{x}{sample ion intensity matrix,row sample, column feature.}

\item{Group}{sample group information}

\item{Batch}{sample batch information. If missing, all the samples will be considered from the same batch}

\item{Seq}{sample sequence with each batch. If missing, the Seq will be automatically assigned according to sample order}

\item{Trans}{How should data be transformed, "LOG2", "LOG10", or NULL transformation?}
}
\value{
a box plot
}
\description{
view variations of TIC among samples
}
\examples{
dat <- matrix(runif(100*9), ncol = 100, nrow = 10)
myGroup <- rep_len(LETTERS[1:5], 10)
mySeq <- c(1:10)
out <- viewTIC(dat, Group = myGroup)
}
