% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewVolcano.R
\name{viewVolcano}
\alias{viewVolcano}
\title{View volcano plot}
\usage{
viewVolcano(result, compare_group, FC = 2, p = 0.05)
}
\arguments{
\item{result}{result from doStat() function}

\item{compare_group}{which groups you want to compare, i.e. c("WT", "Treat1"), only two groups are allowed}

\item{FC}{select fold change values, default = 2}

\item{p}{select p value, default = 0.05}
}
\description{
View volcano plot.
}
\examples{
\dontrun{
viewVolcano(result, compare_group = c("WT", "JA"))
}
}
\author{
Yonghui Dong
}
