% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{search for m/z in from the idiom metabolomics database}
\usage{
what(mz, mode = NULL, ppm = 5, useDB = "HMDB")
}
\arguments{
\item{mz}{m/z values}

\item{mode}{ionization mode, either positive '+' or negative '-'}

\item{ppm}{mass tolerance, default value = 10}

\item{useDB}{which database to use, HMDB or KEGG? default is HMDB}
}
\description{
tentative metabolite identification based on m/z value search
}
\examples{
 a = what(133.014, mode = '-', ppm = 10)
}
\author{
Yonghui Dong
}
