% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaminant.R
\name{contam}
\alias{contam}
\title{Contaminants in MS}
\usage{
contam(mz, mode = NULL, ppm = 10)
}
\arguments{
\item{mz}{suspected m/z value}

\item{mode}{ionization mode, either positive '+' or negative '-'}

\item{ppm}{mass tolerance, default value = 10}
}
\description{
check the possible contaminants
}
\examples{
 contam(33.0335, ppm = 10, mode = '+')
 contam(44.998, ppm = 10, mode = '-')
}
\author{
Yonghui Dong
}
