% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Element_Isotope.R
\name{E_iso}
\alias{E_iso}
\title{Element isotopes}
\usage{
E_iso(S)
}
\arguments{
\item{S}{element, can be element symbol (i.e. C) or full name (i.e. Carbon).
Both Element symbol and full name are case insensitive.}
}
\description{
check element isotope information
}
\examples{
 E_iso('Na') # element symbol
 E_iso('nA') # element symbol, case insensitive
 E_iso('Carbon') # element full name
 E_iso('carBon') # element full name, case insensitive
}
\author{
Yonghui Dong
}
