% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecular_Mass.R
\name{M_mass}
\alias{M_mass}
\title{molecular mass}
\usage{
M_mass(F)
}
\arguments{
\item{F}{chemical formula, case insensitive}
}
\description{
calculate accurate molecular mass
}
\examples{
 M_mass('C7H6O4')
 M_mass('c7H6O4') # case insensitive
}
