% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ion_Mass.R
\name{mz}
\alias{mz}
\title{accurate ion mass}
\usage{
mz(m, z = 1, mode = c("+", "-"))
}
\arguments{
\item{m}{chemical formula of an ion}

\item{z}{charge, default value is 1}

\item{mode}{ionization mode, either positive '+' or negative '-'}
}
\description{
calculate accurate ion mass
}
\examples{
 mz('C7H7O4', z = 1, mode = '+')
}
