% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.R
\name{mz}
\alias{mz}
\title{accurate ion mass}
\usage{
mz(m, z)
}
\arguments{
\item{m}{chemical formula of an ion, case insensitive}

\item{z}{charge}
}
\description{
calculate accurate ion mass
}
\examples{
 mz('C7H7O4', z = 1)
 mz('C10H6Cl1', z = -1)
 mz('C7h7O4', z = 1) # case insensitive
 mz(c('C7H7O4', 'c1'), z = -1) # vector input
}
