% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{mass accuracy}
\usage{
ppm(m, t)
}
\arguments{
\item{m}{measured m/z}

\item{t}{theoretical m/z}
}
\description{
calculate the mass accuracy of measured m/z
}
\examples{
 ppm(155.03383, 155.03388) # with m/z value
 ppm(155.03383, mz('C7H7O4', z = 1, mode = '+')) # with ion formula
}
