% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{search for m/z in from the idiom metabolomics database}
\usage{
what(mz, mode = NULL, ppm = 5)
}
\arguments{
\item{mz}{m/z values}

\item{mode}{ionization mode, either positive '+' or negative '-'}

\item{ppm}{mass tolerance, default value = 10}
}
\description{
tentative metabolite identification based on m/z value search
}
\examples{
 what(133.014, ppm = 10, mode = '-')
 what(c(133.014, 191.020), ppm = 10, mode = '-')
}
\author{
Yonghui Dong
}
